koding/klient: move package to klient that everyone can use it
